requirements.txt
discord
python-dotenv
django
django-allauth
pillow